Move all instance settings in HeadTags to instanceSettings.ts
